# LucasJavaProgramingESpringBoot/JavaSpringBoot/AutoEscolaS164/SwaggerOffLine.yaml
---
openapi: 3.1.0
info:
  title: Auto Escola S-164
  description: Api Rest para gerenciamento de instrução da Auto Escola S-164
  contact:
    name: Auto Escola S-164
    email: autoescola@gmail.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  "/instrutores":
    get:
      tags:
        - instrutor-controller
      operationId: listarInstrutores
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageDadosListagemInstrutor"
    put:
      tags:
        - instrutor-controller
      operationId: atualizarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosAtualizacaoInstrutor"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoInstrutor"
    post:
      tags:
        - instrutor-controller
      operationId: cadastrarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosCadastroInstrutor"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoInstrutor"
  "/alunos":
    get:
      tags:
        - aluno-controller
      operationId: listarAlunos
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageDadosListagemAluno"
    put:
      tags:
        - aluno-controller
      operationId: atualizarAluno
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosAtualizacaoAluno"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoAluno"
    post:
      tags:
        - aluno-controller
      operationId: cadastraAluno
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosCadastroAluno"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoAluno"
  "/Usuario/senha":
    put:
      tags:
        - usuario-controller
      operationId: atualizaSenharUsuario
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosAtualizarSenhaUsuario"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoUsuario"
  "/Usuario/perfil":
    put:
      tags:
        - usuario-controller
      operationId: atualizaPerfilUsuario
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosAtualizarPerfilUsuario"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoUsuario"
  "/login":
    post:
      tags:
        - login-controller
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosLogin"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/dadostokenJWT"
  "/isntrucoes":
    post:
      tags:
        - instrucao-controller
      operationId: agendarInstrucao
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosAgendamento"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentosAgendamento"
  "/Usuario":
    get:
      tags:
        - usuario-controller
      operationId: listarUsuarios
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageDadosListagemUsuario"
    post:
      tags:
        - usuario-controller
      operationId: cadastrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DadosCadastroUsuario"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoUsuario"
  "/instrutores/{id}":
    get:
      tags:
        - instrutor-controller
      operationId: detalharInstutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoInstrutor"
    delete:
      tags:
        - instrutor-controller
      operationId: excluirInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  "/health-check":
    get:
      tags:
        - health-check-controller
      operationId: healthCheck
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/alunos/{id}":
    get:
      tags:
        - aluno-controller
      operationId: detalharAluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoAluno"
    delete:
      tags:
        - aluno-controller
      operationId: excluirAluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  "/Usuario/{id}":
    get:
      tags:
        - usuario-controller
      operationId: detalharUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DadosDetalhamentoUsuario"
    delete:
      tags:
        - usuario-controller
      operationId: excluirUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          "$ref": "#/components/schemas/DadosEndereco"
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
        cep:
          type: string
          minLength: 1
          pattern: "[0-9]{5}-[0-9]{3}"
      required:
        - bairro
        - cep
        - cidade
        - logradouro
        - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          "$ref": "#/components/schemas/Endereco"
        ativo:
          type: boolean
    Endereco:
      type: object
      properties:
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          type: string
        cep:
          type: string
    DadosAtualizacaoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        endereco:
          "$ref": "#/components/schemas/DadosEndereco"
    DadosDetalhamentoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        Telefone:
          type: string
        cpf:
          type: string
        endereco:
          "$ref": "#/components/schemas/Endereco"
        ativo:
          type: boolean
    DadosAtualizarSenhaUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senha:
          type: string
    DadosDetalhamentoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        perfil:
          type: string
          enum:
            - USER
            - ADMIN
        ativo:
          type: boolean
    DadosAtualizarPerfilUsuario:
      type: object
      properties:
        perfil:
          type: string
          enum:
            - USER
            - ADMIN
        id:
          type: integer
          format: int64
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    dadostokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosAgendamento:
      type: object
      properties:
        id_aluno:
          type: integer
          format: int64
        id_isntrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data_hora:
          type: string
          format: date-time
      required:
        - data_hora
        - id_aluno
    DadosDetalhamentosAgendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome_aluno:
          type: string
        nome_instrutor:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data_hora:
          type: string
          format: date-time
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cnh:
          type: string
          minLength: 1
          pattern: "\\d{11}"
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          "$ref": "#/components/schemas/DadosEndereco"
      required:
        - cnh
        - email
        - especialidade
        - nome
        - telefone
    DadosCadastroAluno:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cpf:
          type: string
          minLength: 1
          pattern: "\\d{11}"
        endereco:
          "$ref": "#/components/schemas/DadosEndereco"
      required:
        - cpf
        - email
        - nome
        - telefone
    DadosCadastroUsuario:
      type: object
      properties:
        login:
          type: string
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/DadosListagemInstrutor"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          "$ref": "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    DadosListagemAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
    PageDadosListagemAluno:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/DadosListagemAluno"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    DadosListagemUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
    PageDadosListagemUsuario:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/DadosListagemUsuario"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: beare
      bearerFormat: JWT