# LucasJavaProgramingESpringBoot/JavaSpringBoot/AutoEscolaS164/src/main/java/br/com/senai/autoescolas164/adapter/in/controller/request/usuario/DadosAtualizarSenhaUsuario.java
package br.com.senai.autoescolas164.adapter.in.controller.request.usuario;

public record DadosAtualizarSenhaUsuario(
        Long id,
        String senha
) {
    public String senha(String hash) {
        return hash;
    }
}