# LucasJavaProgramingESpringBoot/JavaSpringBoot/AutoEscolaS164/src/main/java/br/com/senai/autoescolas164/adapter/in/controller/request/usuario/DadosCadastroUsuario.java
package br.com.senai.autoescolas164.adapter.in.controller.request.usuario;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;

public record DadosCadastroUsuario (
    @NotNull
    String login,

    @NotBlank
    String senha

    ){

    public @NotBlank String senha(String hash) {

        return hash;
    }
}