# LucasJavaProgramingESpringBoot/JavaSpringBoot/AutoEscolaS164/src/main/java/br/com/senai/autoescolas164/adapter/in/controller/response/aluno/DadosListagemAluno.java
package br.com.senai.autoescolas164.adapter.in.controller.response.aluno;

import br.com.senai.autoescolas164.aplication.core.domain.Aluno;

public record DadosListagemAluno(
        Long id,
        String nome,
        String email
) {
    public DadosListagemAluno (Aluno aluno){
        this(
                aluno.getId(),
                aluno.getNome(),
                aluno.getEmail()

        );
    }
}